[iOS] refactor: HomeViewModel 가독성 개선
- 본문1

# feat: 새로운 기능 추가
# fix: 버그 수정
# docs: 문서 수정
# test: 테스트 코드 추가
# refactor: 코드 리팩토링
# chore: 빌드 부분 혹은 패키지 매니저 수정사항
# asset: 애셋, 파일 등 추가
