+++ iOS/moti/moti/Presentation/Sources/Presentation/Home/HomeViewModel.swift
@@ -11,71 +11,45 @@ import Core
 import Data
 
 final class HomeViewModel {
-    enum HomeViewModelAction {
-        case launch
-        case addCategory(name: String)
-        case fetchNextPage
-        case fetchCategoryList(category: CategoryItem)
-        case delete(achievementId: Int)
-        case updateAchievement(id: Int, newCategoryId: Int)
-    }
-    
-    enum CategoryListState {
-        case initial
-        case finish
-        case error(message: String)
-    }
-    
-    enum CategoryState {
-        case initial
-        case updated(category: CategoryItem)
-    }
-    
-    enum AddCategoryState {
-        case none
-        case loading
-        case finish
-        case error(message: String)
-    }
-    
-    enum AchievementState {
-        case initial
-        case loading
-        case finish
-        case error(message: String)
-    }
-    
     typealias AchievementDataSource = ListDiffableDataSource<Achievement>
     typealias CategoryDataSource = ListDiffableDataSource<CategoryItem>
     
+    // MARK: - Properties
+    // Category
     private var categoryDataSource: CategoryDataSource?
     private let fetchCategoryListUseCase: FetchCategoryListUseCase
     private let addCategoryUseCase: AddCategoryUseCase
 
-    private var achievementDataSource: AchievementDataSource?
-    private let fetchAchievementListUseCase: FetchAchievementListUseCase
-    
     private var categories: [CategoryItem] = [] {
         didSet {
             categoryDataSource?.update(data: categories)
         }
     }
-    private let skeletonAchievements: [Achievement] = (-20...(-1)).map { Achievement(id: $0, title: "", imageURL: nil) }
-    private var achievements: [Achievement] = [] {
-        didSet {
-            achievementDataSource?.update(data: achievements)
-        }
-    }
-    private var lastRequestNextValue: FetchAchievementListRequestValue?
-    private var nextRequestValue: FetchAchievementListRequestValue?
+    
     private(set) var currentCategory: CategoryItem? {
         didSet {
             guard let currentCategory else { return }
             categoryState = .updated(category: currentCategory)
         }
     }
+    
+    // Achievement
+    private var achievementDataSource: AchievementDataSource?
+    private let fetchAchievementListUseCase: FetchAchievementListUseCase
+    
+    private let skeletonAchievements: [Achievement] = (-20...(-1)).map { Achievement(id: $0, title: "", imageURL: nil) }
+    private var achievements: [Achievement] = [] {
+        didSet {
+            achievementDataSource?.update(data: achievements)
+        }
+    }
+    
+    // Pagenation
+    private var lastRequestNextValue: FetchAchievementListRequestValue?
+    private var nextRequestValue: FetchAchievementListRequestValue?
     private var nextAchievementTask: Task<Void, Never>?
     
+    // State
     @Published private(set) var categoryListState: CategoryListState = .initial
     @Published private(set) var addCategoryState: AddCategoryState = .none
     @Published private(set) var achievementState: AchievementState = .initial
@@ -92,23 +66,7 @@ final class HomeViewModel {
         self.addCategoryUseCase = addCategoryUseCase
     }
     
-    func action(_ action: HomeViewModelAction) {
-        switch action {
-        case .launch:
-            fetchCategories()
-        case .addCategory(let name):
-            addCategory(name: name)
-        case .fetchNextPage:
-            fetchNextAchievementList()
-        case .fetchCategoryList(let category):
-            fetchCategoryAchievementList(category: category)
-        case .delete(let achievementId):
-            delete(achievementId: achievementId)
-        case .updateAchievement(let id, let newCategoryId):
-            updateAchievement(id: id, newCategoryId: newCategoryId)
-        }
-    }
-    
+    // MARK: - Methods
     func setupCategoryDataSource(_ dataSource: CategoryDataSource) {
         self.categoryDataSource = dataSource
     }
@@ -127,28 +85,28 @@ final class HomeViewModel {
         return CategoryStorage.shared.find(categoryId: categoryId)
     }
     
-    private func findIndexOfAchievement(with achievementId: Int) -> Int? {
-        return achievements.firstIndex { $0.id == achievementId }
-    }
-    
-    private func delete(achievementId: Int) {
-        guard let foundIndex = findIndexOfAchievement(with: achievementId) else { return }
-        achievements.remove(at: foundIndex)
-        
-        guard let currentCategoryId = currentCategory?.id else { return }
-        currentCategory = CategoryStorage.shared.find(categoryId: currentCategoryId)
-    }
-    
-    private func updateAchievement(id: Int, newCategoryId: Int) {
-        guard let currentCategory, currentCategory.id != 0 else { return }
-        if currentCategory.id != newCategoryId {
-            CategoryStorage.shared.decrease(categoryId: id)
-            CategoryStorage.shared.increase(categoryId: newCategoryId)
-            delete(achievementId: id)
+    func action(_ action: HomeViewModelAction) {
+        switch action {
+        case .launch:
+            fetchCategories()
+        case .addCategory(let name):
+            addCategory(name: name)
+        case .fetchNextPage:
+            fetchNextAchievementList()
+        case .fetchCategoryList(let category):
+            fetchCategoryAchievementList(category: category)
+        case .delete(let achievementId):
+            delete(achievementId: achievementId)
+        case .updateAchievement(let id, let newCategoryId):
+            updateAchievementCategory(oldCategoryId: id, newCategoryId: newCategoryId)
         }
     }
-    
-    private func fetchCategories() {
+}
+
+// MARK: - Actions
+private extension HomeViewModel {
+    /// 카테고리 리스트를 가져오는 액션
+    func fetchCategories() {
         Task {
             do {
                 categories = try await fetchCategoryListUseCase.execute()
@@ -159,7 +117,8 @@ final class HomeViewModel {
         }
     }
     
-    private func addCategory(name: String) {
+    /// 카테고리 리스트를 추가하는 액션
+    func addCategory(name: String) {
         Task {
             addCategoryState = .loading
             let requestValue = AddCategoryRequestValue(name: name)
@@ -172,8 +131,55 @@ final class HomeViewModel {
             }
         }
     }
+
+    /// 다음 도전 기록 리스트를 가져오는 액션
+    func fetchNextAchievementList() {
+        guard let requestValue = nextRequestValue,
+              lastRequestNextValue?.whereIdLessThan != nextRequestValue?.whereIdLessThan else {
+            Logger.debug("마지막 페이지입니다.")
+            return
+        }
+        lastRequestNextValue = requestValue
+        fetchAchievementList(requestValue: requestValue)
+    }
     
-    private func fetchAchievementList(requestValue: FetchAchievementListRequestValue? = nil) {
+    /// 카테고리의 도전 기록 리스트를 가져오는 액션
+    func fetchCategoryAchievementList(category: CategoryItem) {
+        guard currentCategory != category else {
+            Logger.debug("현재 카테고리입니다.")
+            return
+        }
+        
+        currentCategory = category
+        
+        let requestValue = FetchAchievementListRequestValue(categoryId: category.id, take: nil, whereIdLessThan: nil)
+         fetchAchievementList(requestValue: requestValue)
+    }
+    
+    /// 도전 기록을 삭제하는 액션
+    func delete(achievementId: Int) {
+        guard let foundIndex = firstIndexOf(achievementId: achievementId) else { return }
+        achievements.remove(at: foundIndex)
+        
+        guard let currentCategoryId = currentCategory?.id else { return }
+        currentCategory = CategoryStorage.shared.find(categoryId: currentCategoryId)
+    }
+    
+    /// 도전 기록의 카테고리를 변경하는 액션
+    func updateAchievementCategory(oldCategoryId: Int, newCategoryId: Int) {
+        guard let currentCategory, currentCategory.id != 0 else { return }
+        if currentCategory.id != newCategoryId {
+            CategoryStorage.shared.decrease(categoryId: oldCategoryId)
+            CategoryStorage.shared.increase(categoryId: newCategoryId)
+            delete(achievementId: oldCategoryId)
+        }
+    }
+}
+
+// MARK: - Method
+private extension HomeViewModel {
+    /// 도전 기록을 가져오는 메서드
+    func fetchAchievementList(requestValue: FetchAchievementListRequestValue? = nil) {
         if requestValue?.whereIdLessThan == nil {
             // 새로운 카테고리 데이터를 가져오기 때문에 빈 배열로 초기화
             achievements = skeletonAchievements
@@ -186,44 +192,29 @@ final class HomeViewModel {
             do {
                 achievementState = .loading
                 let (newAchievements, next) = try await fetchAchievementListUseCase.execute(requestValue: requestValue)
-                if let nextAchievementTask, nextAchievementTask.isCancelled {
-                    achievementState = .finish
-                    return
-                }
-                
-                if requestValue?.whereIdLessThan == nil {
+                let isFirstRequest = requestValue?.whereIdLessThan == nil
+                if isFirstRequest {
                     achievements = newAchievements
                 } else {
+                    // 다음 페이지 요청이면 Append
                     achievements.append(contentsOf: newAchievements)
                 }
                 
                 nextRequestValue = next
                 achievementState = .finish
             } catch {
-                achievementState = .error(message: error.localizedDescription)
+                if let nextAchievementTask, nextAchievementTask.isCancelled {
+                    Logger.debug("NextAchievementTask is Cancelled")
+                    achievementState = .finish
+                } else {
+                    achievementState = .error(message: error.localizedDescription)
+                }
             }
         }
     }
     
-    private func fetchCategoryAchievementList(category: CategoryItem) {
-        guard currentCategory != category else {
-            Logger.debug("현재 카테고리입니다.")
-            return
-        }
-        
-        currentCategory = category
-        
-        let requestValue = FetchAchievementListRequestValue(categoryId: category.id, take: nil, whereIdLessThan: nil)
-         fetchAchievementList(requestValue: requestValue)
-    }
-    
-    private func fetchNextAchievementList() {
-        guard let requestValue = nextRequestValue,
-              lastRequestNextValue?.whereIdLessThan != nextRequestValue?.whereIdLessThan else {
-            Logger.debug("마지막 페이지입니다.")
-            return
-        }
-        lastRequestNextValue = requestValue
-        fetchAchievementList(requestValue: requestValue)
+    /// Achievement의 첫 번째 index를 구하는 메서드
+    private func firstIndexOf(achievementId: Int) -> Int? {
+        return achievements.firstIndex { $0.id == achievementId }
     }
 }
